Sync up renering state with reality
Not sure how this got so horrifically out of whack, but our rendered
icons were not matching up with the state they should be in. This
brings everything back together.

Additionally we remove some icons that were only used in Unity. It
should still fall back fairly gracefully if someone is still using
it, but this reduces the amount of stuff we have to ship by a lot.